Added countdown

+++ InteractiveCollages/Views/PhotoView.xaml.cs
@@ -3,33 +3,30 @@ using System.Drawing;
 using System.Drawing.Imaging;
 using System.IO;
 using System.Windows;
+using System.Windows.Controls;
+using System.Windows.Forms.Integration;
 using System.Windows.Media.Imaging;
+using System.Windows.Threading;
 using AForge.Controls;
 using AForge.Video;
 using AForge.Video.DirectShow;
-using UserControl = System.Windows.Controls.UserControl;
-
 
 namespace InteractiveCollages.Views
 {
     /// <summary>
-    /// Interaction logic for PhotoView.xaml
+    ///     Interaction logic for PhotoView.xaml
     /// </summary>
     public partial class PhotoView : UserControl
     {
-        private VideoSourcePlayer videoSourcePlayer;
-        private FilterInfoCollection videoDevices;
-        private VideoCaptureDevice videoDevice;
-        private VideoCapabilities[] videoCapabilities;
+        private readonly int camIndex = 1;
+        private int countdown = 3;
+        private bool hasTakenPic;
         private VideoCapabilities[] snapshotCapabilities;
-        private MainWindow main { get; set; }
+        private VideoCapabilities[] videoCapabilities;
+        private VideoCaptureDevice videoDevice;
+        private FilterInfoCollection videoDevices;
+        private VideoSourcePlayer videoSourcePlayer;
 
-        private bool inPreview { get; set; }
-        private bool hasTakenPic = false;
-        private GreenRemover greenRemover { get; }
-
-
-        private int camIndex = 1;
         public PhotoView(MainWindow main)
         {
             InitializeComponent();
@@ -39,15 +36,59 @@ namespace InteractiveCollages.Views
             camIndex = main.VideoIndex;
             greenRemover.minEffect = main.minGreen;
             greenRemover.maxEffect = main.maxGreen;
+            CreateTimer();
         }
 
+        private MainWindow main { get; }
+
+        private bool inPreview { get; set; }
+        private GreenRemover greenRemover { get; }
+        private DispatcherTimer dispatcherTimer { get; set; }
+
+        private void CreateTimer()
+        {
+            dispatcherTimer = new DispatcherTimer();
+            dispatcherTimer.Tick += dispatcherTimer_Tick;
+            dispatcherTimer.Interval = new TimeSpan(0, 0, 0, 1);
+        }
+
+        private void dispatcherTimer_Tick(object sender, EventArgs e)
+        {
+            if (countdown == 3)
+            {
+                ImageCountdown3.Visibility = Visibility.Visible;
+                countdown--;
+            }
+            else if (countdown == 2)
+            {
+                ImageCountdown2.Visibility = Visibility.Visible;
+                ImageCountdown3.Visibility = Visibility.Hidden;
+                countdown--;
+            }
+            else if (countdown == 1)
+            {
+                ImageCountdown2.Visibility = Visibility.Hidden;
+                ImageCountdown1.Visibility = Visibility.Visible;
+                countdown--;
+            }
+            else
+            {
+                ImageCountdown1.Visibility = Visibility.Hidden;
+
+                dispatcherTimer.Stop();
+
+                if (!inPreview)
+                    while (!hasTakenPic)
+                        CaptureFrame();
+                else if (inPreview) ResetCamera();
+            }
+        }
 
         private void CreateVideoSourcePlayer()
         {
-            
             // Create the interop host control.
-            System.Windows.Forms.Integration.WindowsFormsHost host =
-                new System.Windows.Forms.Integration.WindowsFormsHost();
+            var host =
+                new WindowsFormsHost();
 
             // Create the control.
             videoSourcePlayer = new VideoSourcePlayer();
@@ -59,8 +100,7 @@ namespace InteractiveCollages.Views
 
             // Add the interop host control to the Grid
             // control's collection of child controls.
-            this.GridVideo.Children.Add(host);
-
+            GridVideo.Children.Add(host);
         }
 
         private void StartVideoSourcePlayer()
@@ -72,16 +112,14 @@ namespace InteractiveCollages.Views
 
             if (videoDevice != null)
             {
-                if ((videoCapabilities != null) && (videoCapabilities.Length != 0))
-                {
+                if (videoCapabilities != null && videoCapabilities.Length != 0)
                     videoDevice.VideoResolution = videoCapabilities[camIndex];
-                }
 
-                if ((snapshotCapabilities != null) && (snapshotCapabilities.Length != 0))
+                if (snapshotCapabilities != null && snapshotCapabilities.Length != 0)
                 {
                     videoDevice.ProvideSnapshots = true;
                     videoDevice.SnapshotResolution = snapshotCapabilities[camIndex];
-                    videoDevice.SnapshotFrame += new NewFrameEventHandler(videoDevice_SnapshotFrame);
+                    videoDevice.SnapshotFrame += videoDevice_SnapshotFrame;
                 }
 
                 videoSourcePlayer.VideoSource = videoDevice;
@@ -98,78 +136,70 @@ namespace InteractiveCollages.Views
 
         private void Button_capture_Click(object sender, RoutedEventArgs e)
         {
+            greenRemover.minEffect = main.minGreen;
+            greenRemover.maxEffect = main.maxGreen;
+
             if (!inPreview)
             {
-                while (!hasTakenPic)
-                {
-                    CaptureFrame();
-                }
-
-
+                ButtonCapture.Visibility = Visibility.Hidden;
+                ImageCountdown3.Visibility = Visibility.Visible;
             }
-            else if (inPreview)
-            {
-                ResetCamera();
 
-            }
-            
+            dispatcherTimer.Start();
         }
 
         private void CaptureFrame()
         {
-
-            if ((videoDevice != null) && (videoDevice.ProvideSnapshots))
+            if (videoDevice != null && videoDevice.ProvideSnapshots)
             {
-                bool ready = videoSourcePlayer.InvokeRequired;
+                var ready = videoSourcePlayer.InvokeRequired;
                 if (ready)
-                {
                     videoDevice.SimulateTrigger();
-                }
                 else
-                {
                     videoSourcePlayer.Invoke(new Action(() => videoDevice.SimulateTrigger()));
-                }
             }
         }
+
         private void Capture()
         {
-
             //Takes Photo
 
             //camera.Capture();
 
             //Freezes the camera and show preview of photo
 
-           // Disconnect();
+            // Disconnect();
             //Shows taken picture
-            Bitmap preview = new Bitmap(@"../../Resources/temp/temp.png");
+            var preview = new Bitmap(@"../../Resources/temp/temp.png");
             preview.Dispose();
             //Image_preview.Source = Photo.AsBitmapImage(preview);
 
-            bool ready = UserControl.Dispatcher.CheckAccess();
+            var ready = UserControl.Dispatcher.CheckAccess();
 
             if (ready)
             {
                 Image_preview.Source = greenRemover.RemoveGreen();
-                ButtonCapture.Content = new BitmapImage(new Uri(@"../../Resources/UI/Button_opnieuw.png", UriKind.Relative));
+                ButtonCapture.Content =
+                    new BitmapImage(new Uri(@"../../Resources/UI/Button_opnieuw.png", UriKind.Relative));
                 ImageTevreden.Visibility = Visibility.Visible;
                 ButtonContinue.Visibility = Visibility.Visible;
+                ButtonCapture.Visibility = Visibility.Visible;
                 GridVideo.Visibility = Visibility.Hidden;
-
             }
             else
             {
                 Image_preview.Dispatcher.Invoke(() => { Image_preview.Source = greenRemover.RemoveGreen(); });
-                ButtonCapture.Dispatcher.Invoke(() => { ButtonCapture.Content = new BitmapImage(new Uri(@"../../Resources/UI/Button_opnieuw.png", UriKind.Relative)); });
+                ButtonCapture.Dispatcher.Invoke(() =>
+                {
+                    ButtonCapture.Content =
+                        new BitmapImage(new Uri(@"../../Resources/UI/Button_opnieuw.png", UriKind.Relative));
+                    ButtonCapture.Visibility = Visibility.Visible;
+                });
                 ImageTevreden.Dispatcher.Invoke(() => { ImageTevreden.Visibility = Visibility.Visible; });
+
                 ButtonContinue.Dispatcher.Invoke(() => { ButtonContinue.Visibility = Visibility.Visible; });
-                // videoSourcePlayer.Invoke(videoSourcePlayer.Margin = new Padding(100, 0, 0, 0));
                 GridVideo.Dispatcher.Invoke(() => { GridVideo.Visibility = Visibility.Hidden; });
-
-
             }
-            
-
 
 
             inPreview = true;
@@ -177,26 +207,13 @@ namespace InteractiveCollages.Views
 
         private void ResetCamera()
         {
-            
-
-            //Disables and hides the preview
-            Image_preview.Source = null;
-            inPreview = false;
-
-            //Returns buttons and labels to their original state
-            ButtonCapture.Content = new BitmapImage(new Uri(@"../../Resources/UI/Button_makePhoto.png", UriKind.Relative));
-            ImageTevreden.Visibility = Visibility.Hidden;
-            ButtonContinue.Visibility = Visibility.Hidden;
-
-            
-            Disconnect();
             main.DataContext = new PhotoView(main);
         }
 
         private void ButtonContinue_Click(object sender, RoutedEventArgs e)
         {
             Disconnect();
-            this.Focusable = false;
+            Focusable = false;
             main.DataContext = new CollageView(main);
         }
 
@@ -209,27 +226,19 @@ namespace InteractiveCollages.Views
                 videoSourcePlayer.WaitForStop();
                 videoSourcePlayer.VideoSource = null;
 
-                if (videoDevice.ProvideSnapshots)
-                {
-                    videoDevice.SnapshotFrame -= new NewFrameEventHandler(videoDevice_SnapshotFrame);
-                }
-
+                if (videoDevice.ProvideSnapshots) videoDevice.SnapshotFrame -= videoDevice_SnapshotFrame;
             }
         }
 
         private void videoDevice_SnapshotFrame(object sender, NewFrameEventArgs eventArgs)
         {
             Console.WriteLine(eventArgs.Frame.Size);
-            if (File.Exists(@"../../Resources/temp/temp.png"))
-            {
-                File.Delete(@"../../Resources/temp/temp.png");
-            }
-            Bitmap test = (Bitmap)eventArgs.Frame.Clone();
+            if (File.Exists(@"../../Resources/temp/temp.png")) File.Delete(@"../../Resources/temp/temp.png");
+            var test = (Bitmap) eventArgs.Frame.Clone();
             test.Save(@"../../Resources/temp/temp.png", ImageFormat.Png);
             test.Dispose();
             hasTakenPic = true;
             Capture();
-
         }
 
         private void UserControl_Loaded(object sender, RoutedEventArgs e)
